+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This page shows a detailed overview of the changes between versions without the 
 - Enhancement: Allows using Environment Variables to configure the Matter Server instead of using CLI parameters
 - Enhancement: Displays the global cluster attributes always last in the cluster view of the dashboard
 - Enhancement: Include custom cluster definitions in the generated dashboard data to show attributes and clusters with names
+- Fix: Update matter.js to 0.16.5
+   - Correctly handles already downloaded production OTA updates as production updates 
 
 ## 0.2.3 (2026-01-15)